Filter people from free text
Closes #5

Introduces a dependency on [MITIE Ruby][1]. This commit assumes a file
named "ner_model.dat" lives along side the caller, but a future commit
will make this configurable.

Filters "PEOPLE" from free text so long as the "confidence score" is
above a certain threshold. For now, we pick `0.5`, but a future commit
will make this configurable.

Finally, ignore `.dat` files so that we can test locally.

[1]: https://github.com/ankane/mitie-ruby