Remove `TopSecret::Filter::Regex`
Simplifies implementation in order to better support #5 and #6, since we
don't yet know what that abstraction will look like.

Also, create mapping first in order to optimize our filtering. A future
commit will explore creating a `Mapping` class. Having a mapping will
allow us to not only filter the original text, but any new text that
contains a sensitive word. This will be useful when working with a
Chatbot.