Introduce `TopSecret::Text.filter_all`
Relates to this [discussion][1].

In an effort to filter [conversation state][2] when working with LLMs,
we introduce `TopSecret::Text.filter_all`. This allows us to filter an
Array-like set of messages, while ensuring the mapping is used
consistently throughout each message.

Prior to this commit, you would have needed to call
`TopSecret::Text.filter` on each message. which would have resulted in
inconsistent mapping.

[1]: https://github.com/thoughtbot/top_secret/discussions/47
[2]: https://platform.openai.com/docs/guides/conversation-state?api-mode=responses