Introduce `TopSecret::Text.filter_all` (#54)
Relates to this [discussion][1].

In an effort to filter [conversation state][2] when working with LLMs,
we introduce `TopSecret::Text.filter_all`. This allows us to filter an
Array-like set of messages, while ensuring the mapping is used
consistently throughout each message.

Prior to this commit, you would have needed to call
`TopSecret::Text.filter` on each message. which would have resulted in
inconsistent mapping.

When testing local, I noticed the method was slow. Claude Code wrote some
benchmarks for me, and identified the issue was the fact that we were
initializing `Mitie::NER` for **each** message. After some back-and-forth, we
landed on a solution that introduced dependency injection. This way, we can
initialize `Mitie::NER` just once.

A future commit will introduce a benchmark script and CI step.

[1]: https://github.com/thoughtbot/top_secret/discussions/47
[2]: https://platform.openai.com/docs/guides/conversation-state?api-mode=responses