Add benchmark tooling
While working on #54, I discovered a serious performance issue during
local development. Fortunately I caught it, but worked with Claude Code
to diagnose the problem and implement a solution.

In an effort to catch performance regressions, as well as catch new
performance issues, we introduce `bin/benchmark`

For the time being, this cannot be run in CI because it depends on
`ner_model.dat` existing in root of the project.

It should be noted that `bin/benchmark` will need to be updated each time
a new method is added that we want to test. For now, it will simply
catch regressions.