Introduce `TopSecret::Text::{BatchResult, Result}#{sensitive?, safe?}`
In #70 we introduced `TopSecret::Text#scan`, which returns an object
that responds to `sensitive?`.

This commit updates both `TopSecret::Text::BatchResult` and
`TopSecret::Text::Result` to do the same. In order do do this, we
introduce a `Mapping` module to be shared across all three classes.
We also introduce `#safe?` to compliment `#sensitive?`.

This meant that `TopSecret::Text::BatchResult::Item` had the same
interface as `TopSecret::Text::Result`. Because of this, we remove that
class altogether, and replace it with `TopSecret::Text::Result`.