Cache `Mitie::NER` initialization
Initializing `Mitie::NER` is expensive, so we cache it. In an effort to
make it thread-safe, we use a Mutex.

Although #85 is a simpler approach, it's not practical because if you're
using [Trove](https://github.com/ankane/trove), and [pulling during
asset pre-compilation](https://github.com/ankane/trove), the file will
not exist during initialization.

Claude Code helped drive this out, but it was tested in a
production-like environment and worked. The first request was slow, but
subsequent requests were cached.