Cache `Mitie::NER` (#86)
Initializing `Mitie::NER` is expensive, so we cache it. In an effort to make it
thread-safe, we use a Mutex.

Although #85 is a simpler approach, it's not practical because if you're using
[Trove][1], and [pulling during asset pre-compilation][2], the file will not
exist during initialization.

Claude Code helped drive this out, but it was tested in a production-like
environment and worked. The first request was slow, but sinitialization
ubsequent requests were cached.

[1]: https://github.com/ankane/trove
[2]: https://github.com/ankane/trove?tab=readme-ov-file#heroku-and-dokku